Forbid assigning to free variables.
This makes it impossible (or, at least, very unlikely) to see magic when trying to use local and free variables with the same names in one scope.
(CPython solves this by making all variables one is trying to assign to local unless explicitly stated that a variable is 'nonlocal'. That's almost as bad as changing global variables from a function, though.)