Rename `.each` pseudoattribute to `.for`.
That way it won't clash with any attribute names used by existing Python modules since `for` is a keyword in Python.
Also, `.for` accepts the variable name and the code to evaluate as keyword arguments, meaning it can be called like this:

```
iterable.for each: x do:
  stuff
```

Nice, isn't it? (An optional colon may be added after `for`.)