Add an ability to make compiler extensions.
If you have a module named `yoba`, you can import it as an extension:

  import '/yoba' extension

This will try to import `yoba-macro` (i.e. `yoba-macro.dg` from the same directory as `yoba.dg`), take `infixl`, `infixr`, and `prefix` from its globals, and use them to extend attributes of `Compiler` with the same name. This only affects the code object this is done in, as well as all functions and `where` clauses inside.

To enable an extension globally, use the `-X` option:

  python -m dg -X yoba

An example extension (say, `yoba-macro.dg`):

  prefix = dict'
    'roll', f args ~> @pushconst 'now in cover'

An example program that uses it:

  import '/yoba' extension

  print roll!

Its output:

  now in cover