Allow to replace compiler-defined functions with runtime ones.
Meaning, if the compiler detects there is a local variable with the same name as a compiler-defined function, it will ignore its definition and treat the name as an ordinary variable. For example:

  raise = x -> x
  raise 1  #=> 1 (does not raise any exception)

Exception: if said name has a non-standard symbol scanner, redefinitions are ignored. Names that override the symbol scanner are `for`, `except`, `import`, assignment operators and function definition operators.

  import = x -> x
  import "/sys"  #=> still imports module "sys"

(Partially) fixes #11.