Improve parsing of for/while/with/if/except.
They should now be recognized as the beginning of a block when following any operator equal in strength to the empty operator or weaker, meaning

    a + b for c in d => e

is now parsed as

    a + b (for c in d => e)

and not

    a + b (for) c in d => e   # note that this is actually syntactically invalid

as it was previously. You cannot create a variable named `for` anyway; `for = ...` and `(for) = ...` are both syntactically invalid because the former is parsed as `for (= ...)` and expects a `=>` later, and the latter fails to find a `=>` before a close-paren.